feat: Add pie slice generator
There is now a pie slice coroutine generator which takes a data set and
then returns a generator for corresponding pie slices which can then be
used to make arcs when drawing pie or donut charts.

There is now also an approval test which uses the pie slice generator to
make a donut chart for the main ingredients of an egg & dairy-free
donut.