doc: Add symbol properties documentation
The symbol properties documentation includes a list of the symbols and
their corresponding `char` values which has been added as an approval
test. While creating the approval test a bug was discovered whereby all
line symbols were being rendered in black no matter what color was set.
This is also fixed in this commit.