feat: Rely on aggregate initialization for creation of elements
Until now, the construction of elements was an inconsistent mix of
explicit constructors and aggregate initialization. With the advent of
CTAD for aggregates in C++20, the explicit constructors are no longer
needed. At the moment, however, GCC is the only compiler with support
for CTAD for aggregates. So, for now, *cdv* only compiles on GCC.