feat: Improve facilities for specifying frames
Specify frames using designated initialization so that only the
paramters that the user wants to set actually need to be set.

Allow the specification of margins either in absolute pixels values or
as relative percentages of the frame's total width or height.

Fix issue caused by passing references to `pie_slices` generator (if
temporaries are passed in then they expire before they are actually used
in the coroutine)